+++ hoomd/md/mesh/bending.py
@@ -98,12 +98,12 @@ class Helfrich(MeshPotential):
         \mathrm{Neigh}(i)} \frac{\sigma_{ij}}{l_{ij}} (\mathbf{r}_j
         - \mathbf{r}_i) \right)^2
 
-    with the area of the dual cell of vertex i
+    with the area of the dual cell of vertex :math:`i`
     :math:`\sigma_i=(\sum_{j \in \mathrm{Neigh}(i)}\sigma_{ij}\cdot l_{ij})/4`,
-    the length of the bond in the dual lattice  :math:`\sigma_{ij}=
-    l_{ij}(\text{cot}\theta_1+\text{cot}\theta_2)/2` and the angles
-    :math:`\theta_1` and :math:`\theta_2` opposite to the shared bond of
-    vertex :math:`i` and :math:`j`.
+    the length of the bond between :math:`i` and :math:`j` :math:`l_{ij}`, the length of
+    the bond in the dual lattice  :math:`\sigma_{ij}=l_{ij}(\text{cot}\theta_1+
+    \text{cot}\theta_2)/2` and the angles :math:`\theta_1` and :math:`\theta_2` opposite
+    to the shared bond of vertex :math:`i` and :math:`j`.
 
     See Also:
         * `Gompper and Kroll 1996 <https://doi.org/10.1051/jp1:1996246>`__