+++ CHANGELOG.rst
@@ -17,6 +17,8 @@ Next release
 
 *Fixed*
 
+* Fixed Helfrich equation in the docs
+
 6.1.1 (2026-02-19)
 ^^^^^^^^^^^^^^^^^^^^